- Updated styles for ImageMagnifier - Updated styles for Notifications - Updated types for product - Added useHooks
