+++ public/index.html
@@ -34,7 +34,7 @@
     <title>Store</title>
 </head>
 <body>
-<div id="notifications"></div>
-<div id="root"></div>
+    <div id="notifications"></div>
+    <div id="root"></div>
 </body>
 </html>
