+++ src/features/productDetails/useProductDetails.ts
@@ -1,0 +1,8 @@
+import { useSelector } from "react-redux";
+import { selectProductDetail } from "./productsDetailsSlice";
+
+export const useProductDetails = () => {
+  const { product } = useSelector(selectProductDetail);
+
+  return { product };
+};
