+++ src/routes/index.tsx
@@ -9,7 +9,7 @@ let index: RouteObject[] = [
     path: "/",
     element: (
       <PublicRoute>
-        <ImageMagnifier />
+        <ProductPage />
       </PublicRoute>
     ),
   },
